Moving platform specific modules into new module 'transport'
