+++ src/statemachine.rs
@@ -4,8 +4,7 @@
 
 use crate::consts::PARAMETER_SIZE;
 use crate::errors;
-use crate::platform::device::Device;
-use crate::platform::transaction::Transaction;
+use crate::transport::platform::{device::Device, transaction::Transaction};
 use crate::statecallback::StateCallback;
 use crate::u2fprotocol::{u2f_init_device, u2f_is_keyhandle_valid, u2f_register, u2f_sign};
 use crate::u2ftypes::U2FDevice;
