+++ src/transport/macos/monitor.rs
@@ -5,7 +5,7 @@
 extern crate libc;
 extern crate log;
 
-use crate::platform::iokit::*;
+use crate::transport::platform::iokit::*;
 use crate::util::io_err;
 use core_foundation::base::*;
 use core_foundation::runloop::*;
