+++ src/transport/macos/transaction.rs
@@ -5,8 +5,8 @@
 extern crate libc;
 
 use crate::errors;
-use crate::platform::iokit::{CFRunLoopEntryObserver, IOHIDDeviceRef, SendableRunLoop};
-use crate::platform::monitor::Monitor;
+use crate::transport::platform::iokit::{CFRunLoopEntryObserver, IOHIDDeviceRef, SendableRunLoop};
+use crate::transport::platform::monitor::Monitor;
 use crate::statecallback::StateCallback;
 use core_foundation::runloop::*;
 use std::os::raw::c_void;
