+++ src/transport/stub/device.rs
